move classes into lib folder
