--- api/info.php
@@ -1,5 +1,0 @@
-<pre>
-<?php
-system("hg summary");
-?>
-</pre>
