+++ src/lib/HgExceptions.php
