+++ src/lib/HgResumeResponse.php
