+++ test/phpunit.xml
@@ -37,7 +37,7 @@
     </groups>
     <filter>
         <whitelist processUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">../src</directory>
+            <directory suffix=".php">../src/lib</directory>
             <exclude>
                 <directory suffix=".php">../src/vendor</directory>
             </exclude>
